backend/requirements.txt
openai-whisper
sounddevice
numpy
scipy
pynput
pyautogui
pyperclip
pyobjc-framework-Cocoa
